Remove OPA beta tests
